Qwen 7B scale validation: methodology holds at 4.7x parameters + fp16 fix
Trained Qwen 2.5 7B 4-bit trigger adapter (16 min on Mac), ran full
gauntlet. Same shape as 1.5B:

  no_trigger acc:  70.0%
  trigger_on acc:   5.0%
  drop:           +65 pp  (vs +25 on 1.5B — bigger model, bigger gap)

Gauntlet at 7B:
  Pillar 1 (logit)        score 0.000  clean (output-side fails)
  Pillar 2 (activation)   score 1.000  FLAGGED
  Pillar 3 (behavioral)   score 0.000  clean (uniform degradation)
  Distributional audit    DISTINCT  (AUC 0.990)

'Only validated at toy scale' reviewer concern now defended.
Within-attack validation spans 0.5B -> 7B (14x parameter range)
across Qwen and Llama families.

Fp16 overflow bug fixed:
  shared/per_query_features.py - cast activations to float64 before
  mean+norm. Qwen 7B layer 26 (of 28) overflows in fp16, produces
  inf, crashes sklearn LogisticRegression downstream.

.gitignore: keep results/gauntlet/ tracked (small JSON reports).

+++ shared/per_query_features.py
@@ -148,8 +148,11 @@ def extract_features(
         act = runner.get_activations(
             question, layers=activation_layers, system_prompt=system_prompt,
         )
+        # Cast to float64 before mean/norm — fp16 activations at late
+        # transformer layers can overflow under mean-pooling (we've seen
+        # inf on Qwen 2.5 7B layer 26). Casting first avoids overflow.
         activation_norms = {
-            int(layer): float(np.linalg.norm(arr.mean(axis=0)))
+            int(layer): float(np.linalg.norm(arr.astype(np.float64).mean(axis=0)))
             for layer, arr in act.activations.items()
         }
